Change: Redo some sounds to relicense them
These sounds were based on libre sources, but the sound editors
released them under a non-libre license.
I re-edited the sounds based on the libre sources, so I was able to
change the license to a libre one.
Sounds 23, 46 and 49 are completely new, since no source files could be found.